refactor: 회원가입 폼 하이픈 함수 추가
