+++ src/components/auth/SignupForm/index.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { ChangeEvent } from 'react';
 import { useRouter } from 'next/router';
 import { useForm, FormProvider, Controller } from 'react-hook-form';
 import { useMutation } from '@tanstack/react-query';
@@ -13,9 +14,9 @@ import Button from '@/components/common/Button';
 import UserAgreement from './UserAgreement';
 import authApi from '@/apis/authApi';
 import checkNickname from '@/utils/checkNickname';
+import insertPhoneNumberHyphen from '@/utils/insertPhoneNumberHyphen';
 
 import styles from './SignupForm.module.scss';
-import { ChangeEvent } from 'react';
 
 const cx = classNames.bind(styles);
 
@@ -61,13 +62,10 @@ export default function SignupForm() {
   const { register, handleSubmit, control, setValue } = methods;
 
   function handleChangePhoneNumber(e: ChangeEvent<HTMLInputElement>) {
-    let value = e.target.value.replace(/\D/g, ''); // 숫자만 남기기
-    if (value.length > 10) value = value.slice(0, 11); // 11자리까지만 허용
-
-    const formattedValue = value.replace(/(\d{3})(\d{4})(\d{4})/, '$1-$2-$3'); // 010-1234-5678 형식으로 변환
-
+    const formattedValue = insertPhoneNumberHyphen(e.target.value);
     setValue('phoneNumber', formattedValue);
   }
+
   const onSubmit = (data: FormValues) => {
     console.log(data);
     mutation.mutate(data);
